ai-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
ollama==0.1.7
pypdf==4.0.1
pydantic==2.6.1
python-multipart==0.0.9
httpx==0.25.2
PyMuPDF==1.24.5
pytesseract==0.3.13
Pillow==10.4.0
